Algorithms and Deep Learning Models/Research-Profile-Summarizer/requirements.txt
streamlit
pandas
scholarly
google-generativeai
python-dotenv